Performed end-to-end test
